Support growing and shrinking the hand
Now you can add and subtract three cards at a time.

In the future, it might be nice to do a smarter squash operation when deleting
cards. Such a smart squash could look for blank cards in the remaining hand and
fill them in with the cards that would be deleted otherwise. This would match
the real flow of the game where a set is removed at random from the hand, but
it is not supported yet.